fix values integer for float

+++ src/database/migrations/1643306177877-teste.ts
@@ -14,9 +14,9 @@ export class teste1643306177877 implements MigrationInterface {
         await queryRunner.query(`ALTER TABLE "orders_products" DROP COLUMN "productsId"`);
         await queryRunner.query(`ALTER TABLE "orders_products" ADD "productsId" character varying NOT NULL`);
         await queryRunner.query(`ALTER TABLE "orders_products" DROP COLUMN "unitePrice"`);
-        await queryRunner.query(`ALTER TABLE "orders_products" ADD "unitePrice" integer NOT NULL`);
+        await queryRunner.query(`ALTER TABLE "orders_products" ADD "unitePrice" float NOT NULL`);
         await queryRunner.query(`ALTER TABLE "orders" DROP COLUMN "totalPrice"`);
-        await queryRunner.query(`ALTER TABLE "orders" ADD "totalPrice" integer NOT NULL`);
+        await queryRunner.query(`ALTER TABLE "orders" ADD "totalPrice" float NOT NULL`);
         await queryRunner.query(`ALTER TABLE "orders_products" ADD CONSTRAINT "FK_59a2a83df0dc22d344646cc048f" FOREIGN KEY ("orderOrdersId") REFERENCES "orders"("ordersId") ON DELETE NO ACTION ON UPDATE NO ACTION`);
         await queryRunner.query(`ALTER TABLE "orders_products" ADD CONSTRAINT "FK_4eff63e89274f79195e25c5c115" FOREIGN KEY ("productId") REFERENCES "products"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
         await queryRunner.query(`ALTER TABLE "orders" ADD CONSTRAINT "FK_c6c2785c13b07ddbc4a9792770a" FOREIGN KEY ("usersId") REFERENCES "users"("id") ON DELETE NO ACTION ON UPDATE NO ACTION`);
